refactor: clear session read review blockers
