+++ pkg/services/factory_runtime/internal/services/orchestration/runtime/clean_invocation.go
@@ -11,14 +11,84 @@ import (
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	factory "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/orchestrators/petri"
+	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/rootobservation"
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/scheduler"
 	"github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/state"
 	factorytoken "github.com/portpowered/infinite-you/pkg/services/factory_runtime/internal/services/orchestration/token"
+	"github.com/portpowered/infinite-you/pkg/services/recordings"
 	"github.com/portpowered/infinite-you/pkg/services/work"
 	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
 )
 
 var _ factory.Service = (*factoryImpl)(nil)
+
+func (f *factoryImpl) Observe(ctx context.Context, req factory.ObserveRequest) (factory.ObserveResult, error) {
+	if !validObservationScope(req.Scope) {
+		return factory.ObserveResult{}, factory.ErrInvalidObservationScope
+	}
+	if f == nil {
+		return factory.ObserveResult{}, factory.ErrNotRunning
+	}
+	if ctx != nil {
+		if err := ctx.Err(); err != nil {
+			return factory.ObserveResult{}, err
+		}
+	}
+	f.mu.RLock()
+	state := f.state
+	startedAt := f.startedAt
+	f.mu.RUnlock()
+	switch state {
+	case interfaces.FactoryStateRunning, interfaces.FactoryStatePaused, interfaces.FactoryStateIdle,
+		interfaces.FactoryStateCompleted, interfaces.FactoryStateFailed:
+	default:
+		return factory.ObserveResult{}, factory.ErrNotRunning
+	}
+	if f.engine == nil {
+		return factory.ObserveResult{}, factory.ErrNotRunning
+	}
+	// Runtime observation deliberately reads only the engine-owned detached
+	// boundary. GetEngineStateSnapshot also reconstructs canonical world state
+	// and evaluates enablement for migration-only callers; neither operation is
+	// part of a live status read.
+	snapshot := f.engine.GetRuntimeStateSnapshot()
+	snapshot.FactoryState = string(state)
+	snapshot.Topology = f.topology
+	snapshot.RuntimeStatus = f.deriveRuntimeStatus(state, snapshot)
+	snapshot.LifecycleControlStatus = string(durableLifecycleStatus(state))
+	snapshot.StreamGenerationID = ""
+	if f.eventHistory != nil {
+		snapshot.StreamGenerationID = f.eventHistory.StreamGenerationID()
+	}
+	if !startedAt.IsZero() && f.clock != nil {
+		snapshot.Uptime = f.clock.Now().Sub(startedAt)
+	}
+	result := factory.ObserveResult{Observation: rootobservation.Project(&snapshot, req.Scope)}
+	f.recordRuntimeObservationMetric(req.Scope)
+	return result, nil
+}
+
+const runtimeReadObservationMetricName = "factory_runtime.read.observation"
+
+func (f *factoryImpl) recordRuntimeObservationMetric(scope factory.ObservationScope) {
+	recorder, ok := f.eventHistory.(interface {
+		RecordRuntimeReadMetric(recordings.RuntimeReadMetric)
+	})
+	if !ok || recorder == nil {
+		return
+	}
+	recorder.RecordRuntimeReadMetric(recordings.RuntimeReadMetric{
+		Name: runtimeReadObservationMetricName,
+		Labels: map[string]string{
+			"scope":                    string(scope),
+			"runtime_snapshot_reads":   "1",
+			"operation_count":          "1",
+			"canonical_history_visits": "0",
+			"canonical_events_copied":  "0",
+			"full_history_reductions":  "0",
+		},
+	})
+}
 
 // GetEngineStateSnapshot returns the aggregate observability snapshot for
 // service-facing callers.
