+++ pkg/services/factory_sessions/internal/sessionservice/runtime_metrics.go
@@ -3,7 +3,10 @@ package service
 
 import (
 	factorymetrics "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
+	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
 	sessioninvocation "github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/invocation"
+	"github.com/portpowered/infinite-you/pkg/services/factory_sessions/internal/roles"
+	"github.com/portpowered/infinite-you/pkg/services/recordings"
 )
 
 // Runtime metric names emitted by the transport runtime host.
@@ -72,3 +75,47 @@ const (
 	modelPullMetricFailure       = ModelPullMetricFailure
 	modelPullMetricSourceFailure = ModelPullMetricSourceFailure
 )
+
+type runtimeReadMetricsAdapter struct {
+	recorder roles.InvocationMetricsRecorder
+}
+
+func (adapter runtimeReadMetricsAdapter) record(metric recordings.RuntimeReadMetric) {
+	if adapter.recorder == nil {
+		return
+	}
+	labels := make(map[string]string, len(metric.Labels))
+	for key, value := range metric.Labels {
+		labels[key] = value
+	}
+	adapter.recorder.RecordInvocationMetric(factorysessions.InvocationMetric{
+		Name: metric.Name, Labels: labels,
+	})
+}
+
+func (fs *SessionRuntime) bindRuntimeReadMetrics(bundle factoryRuntimeBundle) {
+	if fs == nil || bundle == nil {
+		return
+	}
+	ledger := runtimeLedgerForReadMetrics(bundle)
+	binder, ok := ledger.(interface {
+		SetRuntimeReadMetricsRecorder(recordings.RuntimeReadMetricsRecorder)
+	})
+	if !ok || binder == nil {
+		return
+	}
+	binder.SetRuntimeReadMetricsRecorder(runtimeReadMetricsAdapter{
+		recorder: fs.invocationMetricsRecorder,
+	}.record)
+}
+
+// runtimeLedgerForReadMetrics keeps optional telemetry compatible with legacy
+// RuntimeRecord test doubles that promote RecordingLedger from a nil embed.
+func runtimeLedgerForReadMetrics(bundle factoryRuntimeBundle) (ledger recordings.Ledger) {
+	defer func() {
+		if recover() != nil {
+			ledger = nil
+		}
+	}()
+	return bundle.RecordingLedger()
+}
